adds writing controller to seperate the logic from the routing
