+++ backend/app/Writings/writingsController.js
@@ -1,12 +1,18 @@
-// might have to require mongoose to work 
-const Writing = require('./writingsModel')
+// might have to require mongoose to work var Writing = require('./writingsModel');
+let Writing = require('./writingsModel');
 
-exports.createWriting = (req, res) => {
+let createWriting = function(req, res) {
+  console.log(req.body);
   const newWriting = new Writing({
       writing: req.body.writing,
   });
-  newWriting.save((err, newWriting) => {
-    if (err) res.send(err);
-    res.json(newWriting);
-  });
-};
+  newWriting.save()
+  .then(item => {
+      res.send("writing saved to database");
+      })
+      .catch(err => {
+      res.status(400).send("unable to save to database");
+      });
+  };
+
+  module.exports={createWriting};
